% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{cl.make}
\alias{cl.make}
\title{Make a command line argument parser}
\usage{
cl.make(name, version, args = commandArgs())
}
\arguments{
\item{name}{The name of the interface}

\item{version}{The version of the interface}

\item{args}{A vector of command line arguments}
}
\value{
A list of functions that parse arguments
}
\description{
Make a command line argument parser
}
