% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{cl.make}
\alias{cl.make}
\title{Make a command line argument parser}
\usage{
cl.make(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{A vector of command line arguments}
}
\value{
A list of functions that parse arguments
}
\description{
Make a command line argument parser
}
\examples{
cl <- cl.make()

n <- cl$int("n", 1, "An integer")
name <- cl$str("name", "Bob", "A name")

if (erratum::is.e(n)) {
  erratum::raise(n)
}

if (erratum::is.e(name)) {
  erratum::raise(m)
}

cat(
  "n =",
  n,
  "name =",
  name,
  "\n"
)
}
